fix: env and infra
