+++ .env.example
@@ -3,4 +3,3 @@ MYSQL_HOST=
 MYSQL_USER=
 MYSQL_PASSWORD=
 MYSQL_DB=
-MYSQL_PORT=
