+++ .env.local
@@ -1,0 +1,6 @@
+PORT=80
+MYSQL_HOST=todolist-deploy-workshop-db-senai.l.aivencloud.com
+MYSQL_USER=pablito
+MYSQL_PASSWORD=1234
+MYSQL_DB=todolist
+MYSQL_PORT=26055
