+++ .gitignore
@@ -1,3 +1,3 @@
-backend/node_modules
-backend/.env
-
+node_modules
+.env
+.env.local
