+++ Dockerfile
@@ -1,0 +1,11 @@
+FROM node:latest 
+
+WORKDIR backend/
+
+COPY ./ . 
+
+RUN npm i 
+
+EXPOSE 80
+
+ENTRYPOINT ["npm", "run", "start"]
