--- backend/.env.example
@@ -1,5 +1,0 @@
-PORT=
-MYSQL_HOST=
-MYSQL_USER=
-MYSQL_PASSWORD=
-MYSQL_DB=
