--- backend/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --env-file=.env src/server.js",
-    "dev": "nodemon --env-file=.env src/server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.18.1",
-    "mysql2": "^2.3.3"
-  },
-  "devDependencies": {
-    "eslint": "^8.22.0",
-    "nodemon": "^2.0.19"
-  }
-}
