+++ src/app.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const cors = require('cors');
+const router = require('./router');
+
+const app = express();
+
+app.use(express.json());
+app.use(cors());
+app.use(router);
+
+module.exports = app;
