+++ src/server.js
@@ -1,0 +1,5 @@
+const app = require('./app');
+
+const PORT = process.env.PORT || 3333;
+
+app.listen(PORT, () => console.log(`Server running or port ${PORT}`));
